quadratic margin for RFP
Bench: 6215754

Elo   | 2.09 +- 1.65 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 2.93 (-2.25, 2.89) [0.00, 3.00]
Games | N: 49630 W: 12608 L: 12310 D: 24712
Penta | [301, 5834, 12263, 6100, 317]
https://openbench.nocturn9x.space/test/6400/

+++ src/cherry/search/search.rs
@@ -380,7 +380,7 @@ pub fn search<Node: NodeType>(
 
     thread.search_stack[ply as usize].static_eval = static_eval;
     if !Node::PV && !in_check && skip_move.is_none() {
-        let rfp_margin = W::rfp_margin(improving, depth);
+        let rfp_margin = W::rfp_margin(improving, depth) as i32;
         if depth < W::rfp_depth() && static_eval - rfp_margin >= beta {
             return if !static_eval.is_win() && !beta.is_win() {
                 Score(static_eval.0 + W::rfp_lerp() * (beta.0 - static_eval.0) / 1024)
